requirements.txt
aiogram>=2.25.1
openai>=0.27.0
python-dotenv>=0.19.0
aiohttp>=3.8.1 